+++ webpack.common.cjs
@@ -29,7 +29,7 @@ module.exports = {
   },
   plugins: [
     new webpack.BannerPlugin({
-      banner: "#!/usr/bin/env node --no-deprecation",
+      banner: "#!/usr/bin/env -S node --no-deprecation",
       raw: true,
     }),
   ],